fix: recompute compare --stat JSON summary for scoped-only/missing findings
Codex review: --format json --stat --used-by/--required-symbol swapped
verdict to the scoped gate result but left summary as stale full-library
counts, since the existing recompute only ran when the payload had a
changes array (to_stat_json omits it). Also splits the post-render
fold-in helpers out of cli_compare_helpers.py into a new sibling module,
cli_compare_fold.py, to stay under the file-size cap.
